app/requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
chromadb
numpy
scikit-learn
tqdm
autogen
sqlite3